requirements.txt
# Core scientific and computation libraries
numpy==1.23.3
scipy

# Audio processing
librosa
soundfile==0.12.1
numba
graiax-silkcoder[libsndfile]

# Natural Language Processing and Text Conversion
unidecode
pyopenjtalk==0.3.2
jamo
pypinyin
pypinyin-dict
jieba
cn2an
inflect
eng_to_ipa
ko_pron
indic_transliteration
num_thai
opencc
fasttext
fastlid
langid
phonemizer==3.2.1
transformers
num2words
mecab-python3
unidic-lite
g2p_en
sentencepiece==0.1.99
jaconv

# Machine Learning and Deep Learning
torch==1.13.0
audonnx
vector_quantize_pytorch

# Web and API services
flask==2.2.3
flask_apscheduler
pydantic==2.3.0
werkzeug==2.3.6
flask_login
flask_wtf

# Compression and Decompression
py7zr

# Other utilities and dependencies
MarkupSafe==2.1.2
six==1.16.0
protobuf==3.20.1
tqdm
PyYAML
py-cpuinfo
cos-python-sdk-v5==1.9.27
python-socketio>=5.11.2
requests
websocket-client